Adjusted the docs files to reflect the naming change. Refactored the hoising of the dataset module
