+++ docs/beta/dataset.md
@@ -1,0 +1,5 @@
+# Dataset
+
+  * [Dataset](/beta/dataset/dataset)
+  * [Record](/beta/dataset/record)
+  * [Dataframe](/beta/dataset/dataframe)
