--- docs/beta/datasets.md
@@ -1,5 +1,0 @@
-# Datasets
-
-  * [Dataset](/beta/datasets/dataset)
-  * [Record](/beta/datasets/record)
-  * [Dataframe](/beta/datasets/dataframe)
