+++ tamr_client/dataset/__init__.py
@@ -1,2 +1,8 @@
-# This __init__.py file is necessary for `mypy --package`
 # See https://github.com/python/mypy/issues/5759
+from tamr_client.dataset.dataset import (
+    from_resource_id,
+)
+from tamr_client.dataset.dataset import Dataset
+from tamr_client.dataset.dataset import DatasetNotFound
+import tamr_client.dataset.record as record
+import tamr_client.dataset.dataframe as dataframe
